Use a wrapped supplier to propagate request scope even for synchronous calls as some FT primitives (like Retry) use executors and may run code in a different thread. Updated new test that fails without this change and passes with it.
